Refactor cover page into its own helper
